requirements.txt
networkx==2.8.8
rdkit-pypi==2022.9.3
torch==1.13.1+cu116
torch-geometric==1.7.2
torch-scatter==2.1.0+pt113cu116
torch-sparse==0.6.15+pt113cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.64.1
ogb==1.3.5
info-nce-pytorch==0.1.4